% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_accounts.R
\name{account_get}
\alias{account_get}
\title{Coinbase API - Accounts - Account Details}
\usage{
account_get(account_uuid = NULL, use_sandbox = FALSE)
}
\arguments{
\item{account_uuid}{Character. Desired account UUID.}

\item{use_sandbox}{Boolean. If TRUE, uses Coinbase sandbox environment. Default: FALSE.}
}
\value{
A list containing account details.
}
\description{
This function queries the \verb{/api/v3/brokerage/accounts/<acc_uuid>} Coinbase API endpoint.
}
\details{
Requires authentication unless \code{use_sandbox = TRUE}.

Retrieves details of the selected brokerage account associated with the authenticated Coinbase
user.

Input validation ensures that malformed or malicious values are rejected early. Use of
this function in production environments should follow best practices for API key management.
}
\examples{
\dontrun{
# Retrieve list of the brokerage account. Use sandbox mode (no authentication required)
account <- account_list("66f975a6-bb2e-44be-82e9-cd8669e404b0", use_sandbox = TRUE)
}
}
\seealso{
\code{\link[=apirequest]{apirequest()}}
}
