% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_addresses.R
\name{address_transactions_list}
\alias{address_transactions_list}
\title{Coinbase API - Onchain Addresses - List Account Address Transactions}
\usage{
address_transactions_list(account_uuid, address_uuid)
}
\arguments{
\item{account_uuid}{A string containing the account UUID.}

\item{address_uuid}{A string containing the address UUID.}
}
\value{
A data.frame of transactions associated with the addresses
}
\description{
This function queries the \code{v2/accounts/{account_uuid}/addresses/{address_uuid}/transactions}
endpoint to retrieve a list of addresses associated with the specified account.
Requires authentication.
}
\examples{
\dontrun{
# List all addresses for an account
tx <- address_transactions_list("c48f20b0-1234-4bd8-90c2-abcde1234567",
                                "12345678-90af-4000-a111-abcdef123456")
}
}
