% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apikey.R
\name{apikey_mem_clean}
\alias{apikey_mem_clean}
\title{API Key - Clean Memory}
\usage{
apikey_mem_clean()
}
\value{
TRUE if no errors happen.
}
\description{
Clean API keys stored in memory.
}
\examples{
apikey_mem_clean()
}
