% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apirequest.R
\name{apirequest}
\alias{apirequest}
\title{API Request - Request}
\usage{
apirequest(
  method = "GET",
  endpoint = NULL,
  path_params = NULL,
  query_params = NULL,
  body_params = NULL,
  need_auth = TRUE
)
}
\arguments{
\item{method}{String. The HTTP method to use (GET, POST, PUT, DELETE, ...). Default: GET.}

\item{endpoint}{String. The endpoint to send the request to.}

\item{path_params}{String. Parameters to be added to the request path. Default: NULL.}

\item{query_params}{Named list. Parameters to be added to the request query. Default: NULL.}

\item{body_params}{Named list. Parameters to be sent as the request body. Default: NULL.}

\item{need_auth}{Boolean. Does this request need authorization? Default: TRUE}
}
\value{
the API response
}
\description{
API Request - Request
}
