% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apirequest.R
\name{apirequest_authorize}
\alias{apirequest_authorize}
\title{API Request - Request Authorize}
\usage{
apirequest_authorize(req)
}
\arguments{
\item{req}{An httr2 request}
}
\value{
The httr2 request received as parameter with an authorization header
}
\description{
API Request - Request Authorize
}
