% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_clock_drift}
\alias{check_clock_drift}
\title{Coinbase API - Utils - Check Time Drift}
\usage{
check_clock_drift(threshold = 30)
}
\arguments{
\item{threshold}{Integer number of seconds allowed before warning is issued. Defaults to 30.}
}
\value{
An invisible object of class \code{rcoinbaseapi_clock_drift}, a list containing:
\itemize{
\item \strong{local:} Local system time (as Unix epoch).
\item \strong{server:} Coinbase server time (as Unix epoch).
\item \strong{time_delta:} Difference between local and server time (in seconds).
}
}
\description{
This function compares your local system clock with Coinbase's public time server,
and warns if a material difference is detected. A significant time drift may cause
authentication errors when interacting with the Coinbase API.
}
\examples{
drift <- check_clock_drift()
print(drift)

}
