% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_public.R
\name{cryptocurrencies}
\alias{cryptocurrencies}
\title{Coinbase API - Public - Cryptocurrencies}
\usage{
cryptocurrencies()
}
\value{
Data Frame
}
\description{
Queries the public \verb{/v3/currencies/crypto} endpoint to get a list of Cryptocurrencies.
}
