% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_public.R
\name{currencies}
\alias{currencies}
\title{Coinbase API - Public - Currencies}
\usage{
currencies()
}
\value{
Data Frame
}
\description{
Queries the public \verb{/v3/currencies} endpoint to get a list of Fiat Currencies.
}
