% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_deposits.R
\name{deposit_confirm}
\alias{deposit_confirm}
\title{Coinbase API - Deposit - Confirm Pending Deposit}
\usage{
deposit_confirm(account_uuid, deposit_uuid)
}
\arguments{
\item{account_uuid}{A string containing the account UUID.}

\item{deposit_uuid}{A string containing the deposit UUID.}
}
\value{
A list with the updated status of the committed deposit.
}
\description{
This function queries the \code{v2/accounts/{account_uuid}/deposits/{deposit_uuid}/commit} endpoint to
manually confirm (commit) a previously created deposit request. Requires authentication.
}
\examples{
\dontrun{
# Confirm a pending deposit
result <- deposit_confirm("c48f20b0-1234-4bd8-90c2-abcde1234567",
                          "fd0f49b4-9876-4f77-bcfe-7e34ad809ad7")
}
}
