% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_deposits.R
\name{deposit_list}
\alias{deposit_list}
\title{Coinbase API - Deposit - List Deposits}
\usage{
deposit_list(account_uuid)
}
\arguments{
\item{account_uuid}{A string containing the account UUID.}
}
\value{
A data.frame of deposit transactions associated with the account.
}
\description{
This function queries the \code{v2/accounts/{account_uuid}/deposits} endpoint to retrieve a list of
deposit transactions associated with the specified account. Requires authentication.
}
\examples{
\dontrun{
# List all deposits for an account
deposits <- deposit_list("c48f20b0-1234-4bd8-90c2-abcde1234567")
}
}
