% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_payment_method.R
\name{pay_method_list}
\alias{pay_method_list}
\title{Coinbase API - Management - List Payment Methods}
\usage{
pay_method_list()
}
\value{
A data.frame with the payment method details.
}
\description{
Coinbase API - Management - List Payment Methods
}
\examples{
\dontrun{
pay_list <- pay_method_list()
}
}
