% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_public.R
\name{server_time}
\alias{server_time}
\title{Coinbase API - Public - Server Time}
\usage{
server_time()
}
\value{
A list containing:
-\strong{epoch}: Server time as Unix timestamp (seconds since epoch).
-\strong{iso}: Server time in ISO 8601 format (UTC).
}
\description{
Queries the public \verb{/v2/time} endpoint to retrieve the current Coinbase server time.
}
\examples{
server_time()

}
