% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_transactions.R
\name{transaction_list}
\alias{transaction_list}
\title{Coinbase API - Transactions - List Transactions}
\usage{
transaction_list(account_uuid)
}
\arguments{
\item{account_uuid}{Character. Desired account UUID.}
}
\value{
A data frame containing transaction details.
}
\description{
This function queries the \verb{/v2/accounts/<accid>/transactions} Coinbase API endpoint. Requires
authentication.
}
\details{
Retrieves a list of transactions associated with the account passed as parameter.

Input validation ensures that malformed or malicious values are rejected early. Use of
this function in production environments should follow best practices for API key management.
}
\examples{
\dontrun{
# Retrieve list of all transactions for the account
accounts <- transaction_list("00000000-0000-4000-8000-000000000000")
}
}
