% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_transactions.R
\name{transaction_send}
\alias{transaction_send}
\title{Coinbase API - Transactions - Send Cryptocurrency}
\usage{
transaction_send(account_uuid, to, amount, currency)
}
\arguments{
\item{account_uuid}{Character. Desired account UUID.}

\item{to}{Character. A blockchain address, or email of the recipient.}

\item{amount}{Numeric. Amount to be sent.}

\item{currency}{Character. Cryptocurrency of the amount.}
}
\value{
List with transaction details.
}
\description{
Coinbase API - Transactions - Send Cryptocurrency
}
\examples{
\dontrun{
tx <- transaction_send("00000000-0000-4000-8000-000000000000",
                       "3JA7MtC4eXMCWsgJJr9eUCLP2twSy4ouJn",
                       0.0000001,
                       "BTC")
}
}
