% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_transactions.R
\name{transaction_summary}
\alias{transaction_summary}
\title{Coinbase API - Transactions - Transactions Summary}
\usage{
transaction_summary(product_type = NULL, expiry = NULL, venue = NULL)
}
\arguments{
\item{product_type}{Optional. One of \code{"UNKNOWN_PRODUCT_TYPE"}, \code{"SPOT"}, or \code{"FUTURE"}.}

\item{expiry}{Optional. One of \code{"UNKNOWN_CONTRACT_EXPIRY_TYPE"}, \code{"EXPIRING"}, or \code{"PERPETUAL"}.
Only applicable when \code{product_type = "FUTURE"}.}

\item{venue}{Optional. One of \code{"UNKNOWN_VENUE_TYPE"}, \code{"CBE"}, \code{"FCM"}, or \code{"INTX"}.}
}
\value{
A named list containing at least the following:
\itemize{
\item \strong{total_volume}: Total volume across assets, denoted in USD.
\item \strong{total_fees}: Total fees across assets, denoted in USD.
\item \strong{fee_tier}: A list with the description of maker and taker rates across all products.
}
}
\description{
Retrieves a summary of user transactions from the
\verb{/api/v3/brokerage/transaction_summary} endpoint. Authentication is required.
}
\details{
The response includes total trading volume, total fees paid, and maker/taker fee tiers.
}
\examples{
\dontrun{
# Retrieve list of transactions for the account
tx_summary <- transaction_summary()
}
}
