% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_withdrawals.R
\name{withdraw_list}
\alias{withdraw_list}
\title{Coinbase API - Withdraw - List Withdrawals}
\usage{
withdraw_list(account_uuid)
}
\arguments{
\item{account_uuid}{A string containing the account UUID.}
}
\value{
A data.frame of withdraw transactions associated with the account.
}
\description{
This function queries the \code{v2/accounts/{account_uuid}/withdrawals} endpoint to retrieve a list of
withdraw transactions associated with the specified account. Requires authentication.
}
\examples{
\dontrun{
# List all withdraw for an account
withdraw <- withdraw_list("c48f20b0-1234-4bd8-90c2-abcde1234567")
}
}
