% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apikey.R
\name{apikey_read}
\alias{apikey_read}
\title{API Key - Read}
\usage{
apikey_read(password = NULL, timeout = 60L)
}
\arguments{
\item{password}{String. The password used to encrypt the key.}

\item{timeout}{Integer (default = 60s). Time, in seconds, to keep the keys loaded in memory. If
0 (zero) is provided this protection will not be active.}
}
\value{
Stores key information in memory and returns TRUE if successful.
}
\description{
Reads and decrypts a previously stored API key.
}
\examples{
\dontrun{
apikey_read("a_very_strong_password_that_nobody_will_ever_find")
}

}
