% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key.R
\name{apikey_read}
\alias{apikey_read}
\title{API Key - Read}
\usage{
apikey_read(password = NULL)
}
\arguments{
\item{password}{String. The password used to encrypt the key.}
}
\value{
A list with elements \code{name} and \code{key}, or stops with an error on failure.
}
\description{
Reads and decrypts a previously stored API key.
}
\examples{
\dontrun{
apikey_read("a_very_strong_password_that_nobody_will_ever_find")
}

}
