% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apikey.R
\name{apikey_store}
\alias{apikey_store}
\title{API Key - Store}
\usage{
apikey_store(
  key_name,
  key,
  password = NULL,
  overwrite = FALSE,
  key_path = rappdirs::user_config_dir(utils::packageName())
)
}
\arguments{
\item{key_name}{String. The key name.}

\item{key}{String. The API key.}

\item{password}{String. A password to encrypt the key for storage.}

\item{overwrite}{Boolean. What to do if a previous key file is detected and function is called in
non-interactive session.}

\item{key_path}{String. Path where to store the key file. By default it will pick a directory
on user's configuration folder.}
}
\value{
Silently, the key file path if successful.
}
\description{
Stores the API key securely for latter retrieval.
By default it will not overwrite a previously existing key unless explicitly set.
}
\examples{
\dontrun{
# WARNING: running this example may result in a previously stored key to be overwritten.
apikey_store("my_key_name",
             openssl::ec_keygen(),
             "a_very_strong_password_that_nobody_will_ever_find")
}

}
