% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apirequest.R
\name{apirequest}
\alias{apirequest}
\title{API Request - Request}
\usage{
apirequest(method = "GET", endpoint = NULL, need_auth = TRUE)
}
\arguments{
\item{method}{HTTP method to use. Defaults to GET}

\item{endpoint}{The endpoint to send the request to}

\item{need_auth}{Does this request need authorization?}
}
\value{
the API response
}
\description{
API Request - Request
}
