% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_public.R
\name{server_time}
\alias{server_time}
\title{Coinbase API - Public - Server Time}
\usage{
server_time()
}
\value{
A list containing:
-\strong{iso}: Server time in ISO 8601 format (UTC).
-\strong{epochSeconds}: Server time as Unix timestamp (seconds since epoch).
-\strong{epochMillis}: Server time as Unix timestamp (milliseconds since epoch).
}
\description{
Queries the public \verb{/api/v3/brokerage/time} endpoint to get the current Coinbase server time.
}
\examples{
server_time()

}
